Move ContactRelation class to Contact\Relation
- Update references in GContact
